refactor : CacheConfig 최종 설정
